+++ CHANGELOG.md
@@ -37,6 +37,7 @@
 * Sort pieces on bind items screen by expected receipt date. Refs UIREC-392.
 * Leverage new endpoint to get items across tenants. Refs UIREC-388.
 * *BREAKING* Bump up okapi interfaces for `pieces` (3.1) to support a new endpoint. Refs UIREC-396.
+* Only Display pieces on receiving title view if user affiliation(s) matches piece location affiliation. Refs UIREC-390.
 
 ## [5.0.5](https://github.com/folio-org/ui-receiving/tree/v5.0.5) (2024-08-05)
 [Full Changelog](https://github.com/folio-org/ui-receiving/compare/v5.0.4...v5.0.5)